Telephony: Fix to show both voice and data service states
To display service state under phone settings, only voice registration state
value is considered. This causes confusion when data is registered
on the device and signal strength icon is shown but voice will be out of service.

Fix is to display both voice and data registration states under "Service state"
Bug-Id: 38446847

Change-Id: Ibdafd90467f1baa80ce2d871e4d466e68b5ca286