Bluetooth: PBAP: Add support for PBAP UI preference
Usecase:
1) Connect with carkit that supports PBAP
2) Try to uncheck the checkbox for contact sharing from device sub settings menu

Expected Result
1) Should be able to disconnect when checkbox moves from selected->unselected,
and carkit should be able to connect when checkbox is selected again

Observation
Nothing happens. PBAP connection remains active and if in beginning checkbox is not
selected, PBAP cannot be connected.

Fix:
Add support for PBAP UI preference in SettingsLib.

Test:Connect from carkit, try disconnecting from sub settings menu by unchecking
Contact Sharing checkbox and check if PBAP is successfully disconnected.

Bug: 35014213
Change-Id: I4981aa063b2541b58ce1d36e3576578cbeb02acc