Fix two reset_network swedish translations
The reset_network_title translation meant "Network settings HAS BEEN
restored" which is confusing as it is a command to restore network
settings.

The reset_network_complete_toast translation meant "Network settings
has NOT been restored" which is wrong since it is used to indicate
that the settings actually were restored.

Bug: 64291820
Test: manual
Change-Id: I9f2feb0f1ef425b292f2f94bc8482494df430413