Update patch set 2
Patch Set 2:

We won't be supporting priority for EXTRA_SETTINGS_ACTION directly, but we do keep OEM customization in mind when working on things like this for Settings and this will be easier in the future.

There is an easy way to place where extra settings fall within each category, but that only allows one spot per category.

Patch-set: 2