Update patch set 2
Patch Set 2:

Thank you for your reply.
"This will be easier in the future" is good information for us.
Some operators request us to place their settings at the top of category or dashboard.
So we will need priority (customization) for OPERATOR_SETTINGS.
Do you have any plans to support it?

Patch-set: 2