Update patch set 3
Patch Set 3: Code-Review-2

Wow.  I'm not sure which fix I like better.

This one seems to suggest that setValue clobbers the summary.  Is this new behavior?  The documentation for ListPreference doesn't mention it.

The other change suggests what's happening is setValue causes onSharedPreferenceChanged to be called which is zapping the summary that was set.

I think I prefer the other fix simply because it reduces the cpu load: eliminate the second setting of summary (and the lookup of what to set it to.  It was close though and I thank both of you for the research and though you put in.

Patch-set: 3
Reviewer: Robert Greenwalt <1002609@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2