Update patch set 1
Patch Set 1:

It's not currently a mode though, only a hidden options panel. Lots of people that aren't developers make use of it. In some regions like China, sideloading apps via adb is extremely common.

It would be better to expose the property to the shell user, and have native profiling tools automatically enable it. Still a reduction in security, but much less of one than enabling it simply because the hidden developer options panel is enabled.

Patch-set: 1