Update patch set 1
Patch Set 1:

The correct fix here is for Vpns.get(userId) to not be null.

In MNC the service was (wrongly) listening to USER_STARTING and USER_STOPPING which are background broadcasts. Changing it to USER_STARTED and USER_STOPPED seems to fix it.

Does this reproduce on the NYC previews?

Patch-set: 1