Update patch set 1
Patch Set 1:

Hi Robin, I'm not sure if this can reproduce on the NYC previews.  Because we're running monkey test to find this issue.  And if you want, I can find a nexus to try it. 

I think, if we can't fix this issue in the service completely, then change it to Exception maybe a not bad idea.

Patch-set: 1