Update patch set 1
Abandoned

> Thanks for the patch. Unfortunately, I'm unable to reproduce the
 > problem on my build of master. Actually, step 3 shouldn't be
 > possible unless there's a bug with refreshing fingerprints on the
 > device in question. I think that's the source of this bug.

Thanks for the review Jim. You were right. There was a driver bug causing delays in refreshing fingerprints. The onRemoved callback is still async with the removal of the fragment so in theory I still think the change is valid. In practice it seems exceedingly unlikely that one would trigger it with a sane driver. Abandon.

Patch-set: 1
Status: abandoned