Broadcast intent ACTION_MOBILE_DATA_TOGGLE when toggling mobile data
Broadcast the new intent ACTION_MOBILE_DATA_TOGGLE when toggling mobile
data in Settings instead of showing the dialogs directly in Settings.
This ensures that the dialogs that is shown when toggling mobile data is
consistent regardless of where the user toggle mobile data. There are
today multiple different implementations with slightly different dialogs
that may be shown depending on where the user toggled mobile data.

This change is performed in the following locations in Settings:
-MobileDataPreferenceController:
Settings..Network & Internet..Mobile network

-CellularDataConditionController:
Settings..Toggle in the "Mobile Data is off"-conditional

-CellDataPreference:
Settings..Tap the "Mobile Data is off"-conditional

Guest user do not have the neccessary permissions to broadcast the
intent ACTION_MOBILE_DATA_TOGGLE. The mobile data toggle is therefore
disabled for Guest users in the screens above.

Bug: 117861767
Test: make RunSettingsRoboTests ROBOTEST_FILTER=DataPreference
Change-Id: I3ce2d49756339d8d35e858f511d56c164f03390e