Hide cellular network signal in battery history detail when the device is wifi only
It's no sense showing cellular network signal when the device is pad. We should hide
the item when the device is wifi only.Hide cellular network signal when the device is wifi only

It's no sense showing cellular network signal in battery history
detail when the device is pad. We should hide the item when the
device is wifi only.

Bug: https://code.google.com/p/android/issues/detail?id=233639

Test: manual-go to battery history detail and observe the view.

Change-Id: I91f456968268c366741ad16977d3ab0f14725ce5
Signed-off-by: gaochong <gaochong@xiaomi.com>