Update patch set 8
Patch Set 8: Code-Review-2

I think that onPause may not be the right part of the activity life cycle here. I feel like onPause could be called if the activity is simply interrupted by another dialog that pops in front - it is NOT simply a response to the back button. The back button can (and should) be handled differently.

Patch-set: 8
Reviewer: Andre Eisenbach <1058798@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2