Show security info in saved network list
When an access point linked to a saved network is not found,
security info should be shown in dialog of saved network.
Notes that in Android N this list is created with saved networks
and scan result, and that causes security info is not shown for
access points seen around user, but in Android O only saved
networks are used to create a list. That means security info is
shown for all access points listed in the screen.

Bug: 38102896

Change-Id: Ia0c5184ae7a1325f2628f4c08e29350c5793a08f