Update patch set 2
Patch Set 2:

Although settings has no primary use, I'm in no rush to remove it, as it's a useful debugging tool, and allows us to view the state of WifiP2p.
I think it should eventually be removed, but now is not the time.
Is there any reason why this must go in now?

Patch-set: 2
Reviewer: Glen Kuhne <1094298@85c56323-6fa9-3386-8a01-6480fb634889>