Fix error toast isn't readable when wrong pin code
The error toast is displayed when user inputs wrong PIN code in three
times, then the popup to require SIM PUK code will come up and cover
the toast.
As a result, the user can't seen the error toast.

To resolve it, create a customzied view which can be displayed over PUK
unlock screen.

bug: 117894336
Change-Id: I569a2fc2adac15c4713b4c95e9720335bd4df88a