Update patch set 3
Patch Set 3:

> At query time if user types katakana the string will happily match, 
> but what if the user still types hiragana?

The query string is also converted to katakana, so even if user types hiragana, the string will be converted to katakana.
The disadvantage of this solution is that we cannot tell hiragana apart. But I think that this solution has more advantages than disadvantages.

> also, I guess kanji is entirely out of scope?

Yes, kanji is entirely out of scope.

Patch-set: 3
Reviewer: Yoshinori Hirano <1077031@85c56323-6fa9-3386-8a01-6480fb634889>