Update patch set 3
Patch Set 3: Code-Review+1

Hey Yoshinori!

This looks good. We're in the middle of some refactoring of these classes, which means that this CL will cause some pretty big merge conflicts.

If it's okay with you, I will take your code and move it to the new location internally, so the changes will be ready in P.

Patch-set: 3
Reviewer: Matthew Fritze <1125084@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1