Update patch set 2
Patch Set 2:

> (1 comment)

Hello Doris, "if the app is uninstalled, it will not crash" refers to the second case shown in Bug link，and the real crash happens in AppStateAppOpsBridge rather than UsageAccessDetails. 
I find you have fixed it and added the testcase in AppStateAppOpsBridgeTest called getPermissionInfo_nullPackageInfo_shouldNotCrash. 
Do you mean add a testcase for the first case: if an invalid packageName is passed to AppInfoBase, UsageAccessDetails will not crash in refreshUI ?

Patch-set: 2
Reviewer: siming tian <1143183@85c56323-6fa9-3386-8a01-6480fb634889>