Preference SimSettingsActivity not aligned correctly
The preference SimSettingsActivity in the screen "Network & Internet"
never uses a summary. However the text for this preference ("SIM cards")
is top aligned in order to make room for an extra line for the
non-existing summary.

Solved by setting the summary to @string/summary_empty for the
preference SimSettingsActivity. This will make the text in the
preference to align correctly and not leave extra room for a summary.

Test: manual - Checked that preference SimSettingsActivity is correctly
aligned in screen "Network & Internet".
Bug: 69405185
Change-Id: I1a1b3fa7253bbf2a5299f54cc8c4d3e857ca059c