The text on toast should be changed when remains 1 time for SIM lock input
The behavior of getQuantityString() is different according to language, it gets
both "other" and "one" string if English is system language. But it always gets
"other" string if non-English, then the "other" string of wrong_pin_code is
shown even if remains 1 time of SIM lock input.

In order to change the string in any language, we add new string for
remains 1 time.

Bug: 71783850
Change-Id: I5de35eb71905b77028bf25226381e1ba79a37e92