Update patch set 1
Patch Set 1:

(3 comments)

Can you clarify, what is the user clicking when the app crashes?

I'm about worried about this solution - I'm not sure that catching the crash is the right thing to do here. Even if the app doesn't crash, the user is left with a view that becomes a no-op, which is confusing and at worst very frustrating (if that's the operation they wanted to do).

Can we instead disable or hide the view that crashes Settings?

Patch-set: 1
Reviewer: Matthew Fritze <1125084@85c56323-6fa9-3386-8a01-6480fb634889>