Update patch set 2
Patch Set 2:

Pairing state machine had already handled this. On bond state NONE we are dismissing the pairing pop up, but here it is a race condition. In bluetoothPairingDialog.java at Oncreate, after showing the pairing pop up we are registering for the intent ACTION_BOND_STATE. Here the 
pairing dialog is taking more time(100-200ms) and within that time frame we are getting ACTION_BOND_STATE_CHANGED intent but we are not listening for this intent at BluetoothPairingDialog.java

Patch-set: 2
Reviewer: Srinu Jella <1094383@85c56323-6fa9-3386-8a01-6480fb634889>