Use the actionbar themed context for search text
The search textview's color is set to the primary text color. However,
it uses the activity context instead of the actionbar themed context
which works fine for both light/dark colorschemes but does not work for
a light theme with a dark actionbar.

This patch uses the actionbar themed context instead to resolve the
color, which makes using RRO with a DarkActionbar theme possible.

Change-Id: I40646e606f4a8020979534a239af9c962c94548e