MSIM: Fix preferred apn issue for same operator sims
On a multisim device, if same operator sims are used, same preferred apn
is displayed for both subs. When a preferred apn is set for one sub,
this preferred apn is displayed for both subs, since same db entry is
accessed for display.

Fix is to unselect apn if it doesnt match preferred apn criteria in
ApnSettings

Bug: 80457193
Change-Id: If738c8b6ed33b1f48321a0cd4f17487998197f5d