Fix an invalid input value bug in Data warning.
Insert SIM card, In Settings->Network & Internet
->Data usage ->Billing cycle->Data warning, if user
just input a character "." and click SET, Settings
crash. Input type of EditText in xml has defined to
numberDecimal, which not allow to input two "."
characters and blank, but allow start with "." and
end with ".",refer the way to handle invalid input
value, set "." to "0".

Test: Input "." in Data warning and click SET, settings not crash.
Change-Id: I180c397740b9494a97f8bd6467040c83cda0849e