Modify dialog wording for partial connectivity feature
In some areas(e.g. China), Network validation may only pass http
validation but fail to https validation. For this kind of
network, it may still work at some websites or apps. In this
situation, device will popup a notification and dialog for
user to decide if he/she still want to use this network or not.

Test: 1. Build pass.
      2. runtest frameworks-net
      3. Fake partial connectivity case for testing.

Change-Id: I8de323c7c5f356c28a08627cd31e8098f3a22c5d