Handle the intent params from storage notification properly
After migration of private volume to internal storage, notification is
shown to guide user to the settings screen of internal storage.

When the notification is tapped, PrivateVolumeSettings is launched
through PrivateVolumeSettingsActivity with volume id as extra info,
but PrivateVolumeSettings cannot retrieve the extra info and results in
Settings app crash.

This CL fixes to handle the extra info in PrivateVolumeSettings
when launched from PrivateVolumeSettingsActivity.

Also fix the PrivateVolumeForget, which has the same issue as
PrivateVolumeSettings.

Bug: xxx
Test: Settings robotests & manual
Change-Id: I3c5c155155919930fa4474b70b64eb3c867697d3