Update patch set 1
Patch Set 1: -Verified -Code-Review

We should probably add a listener for ImsConfig changes here as well. I modify the provisioning state manually through shell commands and this UI behaves weirdly if the device changes provisioning state while the UI is open.

For example, if the WFC state changes from provisioned-> not provisioned, the WFC option is still available and if you click on it and then try to enable WFC, the text changes to "On", but the slider stays off...

We should implement the same ProvisioningManager callback mechanism here as the one in p/s/Telephony to update the menu when provisioning status changes.

Patch-set: 1
Reviewer: Gerrit User 1098053 <1098053@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0
Label: Code-Review=0