Update patch set 3
Patch Set 3: Code-Review-2

I don't think a developer setting is the right thing. Developer settings is not for us the developers of Android, it's for people who develop ON android. This is meant to turn on logs in logcat, so please have an ADB command to do this instead.

Patch-set: 3
Reviewer: Gerrit User 1015428 <1015428@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2