Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Is it clean cherry pick without any manual change?

Yes, this change is clean. The follow up change which makes the dashboard is not that clean, because of missing settings:searchable in xml and the @SearchIndexable annotation. But the main functionality is surprisingly working well, except the default layout of BasePreferenceController for list preference has some issue. I think I could tweak to make it work.

Patch-set: 3
Reviewer: Gerrit User 1339259 <1339259@85c56323-6fa9-3386-8a01-6480fb634889>