Allow SimDialogActivity to be started for result
In Telephony, we sometimes need to be able to launch
the SIM pick dialog activity for SMS without setting
a new default (i.e. user chose "ask every time") for
default SMS subscription.

Adds the ability for the SimDialogActivity to be
called using startActivityForResult and not set the
user's choice as the default.

Bug: 130853716
Bug: 130567323
Test: run Telephony SmsManagerTestApp
Merged-In: If9f9ebbfe9b7b6718ed759937abbcfa6d22c6295
Change-Id: If9f9ebbfe9b7b6718ed759937abbcfa6d22c6295