Update patch set 1
Patch Set 1:

FYI, if you look at the latest Android Q beta releases, you'll see that this UI got changed from a dialog to a regular settings screen, so this fix would need to be reworked for the new code for that (I'm not sure off the top of my head if the source for the public Q betas has been released yet). 

Backup up a step, can you explain a bit more about why this change is needed? Is there some problem with displaying the raw baseband string even if it does accidentally contain a comma-separated list of duplicate strings? Is this something that can be seen on multiple devices from various OEMs, or is one OEM doing something unusual?

Patch-set: 1
Reviewer: Gerrit User 1214924 <1214924@85c56323-6fa9-3386-8a01-6480fb634889>