MSIM: Fix preferred apn issue for same operator sims
On a multisim device, if same operator sims are used, same preferred apn
is displayed for both subs. When a preferred apn is set for DDS sub,
though preferred apn is null for non DDS sub, since same db entry is
accessed for display, it is displayed as preferred for non DDS sub.

Fix is to unselect apn if it doesnt match preferred apn criteria in
ApnSettings

Bug: 145945660
Change-Id: I2d77572f72d1374126aca43ae1135486715b5198