Add camera button long-press preference
Allow chosing to launch a camera application by long-pressing the
physical camera button on devices that have one.

A new setting will appear in System -> Gestures.

Depends on change in frameworks/base and on a bool in the framework
config named "config_cameraButtonLaunchEnabled".

Same as DoubleTapPowerPreference, the VideoPreference is not implemented
for now.

Change-Id: Ic07b55edb81b557a264cfab8972a6f742a43eca4
Signed-off-by: Felix <google@ix5.org>