Add <uses-library> information.
It is necessary to add the required <uses-library>
"org.apache.http.legacy" explicitly to `uses_libs`, because it is not
among the app dependencies (so Soong cannot add it automatically to the
manifest_fixer), and also because since https://r.android.com/1424793
Soong considers "org.apache.http.legacy" optional (so it would add it to
`optional_uses_libs` instead of `uses_libs`). Soong also doesn't add it
as a compatibility library because the target API level for the app is
not below 28.

Bug: 132357300
Test: lunch cf_x86_phone-userdebug && m
Test: m out/soong/.intermediates/packages/apps/Settings/Settings/android_common/manifest_check/AndroidManifest.xml
Test: m out/soong/.intermediates/packages/apps/Settings/tests/robotests/SettingsRoboTestStub/android_common/manifest_check/AndroidManifest.xm
Change-Id: Id5d5fe1dff68cf6fce667a00e35710cf0e667d02