Add missing null check for read-only apn types
Problem:
Create a new apn without an apn type specified. When the apn is
saved, it causes the Settings app to crash.

Rootcause:
While populating the apn type to all non-readonly types, the logic
is missing a null check when it tries to convert the array mReadOnlyApnTypes
into a list. If the array is null, it causes the Settings app to crash.

Fix:
Add a suitable check for null array.

Change-Id: If9b47b5f79419c7a400dd70bfb40772ce680212c