fix resultCodes from WifiPickerActivity
The current flow is that the android.net.wifi.PICK_WIFI_NETWORK Intent
is received by WifiPickerActivity, and then

* WifiPickerActivity launches the WifiSettings Fragment,
* WifiSettings launches WifiSettings2Activity (due to the
  SETTINGS_WIFITRACKER2 feature flag), and finally
* WifiSettings finishes the Activity it is attached to
  (WifiPickerActivity).

Instead of relying on WifiSettings to launch WifiSettings2, we should
just launch WifiSettings2 directly. If we don't do this, WifiSettings
finishing WifiPickerActivity will send a resultCode of RESULT_CANCELLED
to the app that launched the WifiPickerActivity.

This means that the resultCode from WifiSettings2 will never get
propagated, so apps that use things such as extra_prefs_show_button_bar
(like setup wizards) may not work correctly if they rely on the
resultCode.

Test: manual, using an app with an Activity that launches
WifiPickerActivity and then prints out the resultCode. It's launched
using

    val intent = Intent(WifiManager.ACTION_PICK_WIFI_NETWORK)
        .putExtra("extra_prefs_show_button_bar", true)
    startActivityForResult(intent, WIFI_PICKER_REQUEST_CODE)

Change-Id: I1253867bd54a4d231e5aaf6e295bdde5dfddedd2