Disable resource deduping
Settings app lacks wifi_calling_settings_title resource
for values-mcc262-mnc02.

aapt2 resource deduping removes mcc mnc
specific values when those are the same as the default value.
This causes problems when the carrier has choose to set a
value that overrides the language default.

This Bug was found when language is set to German and you
use a Vodafone DE sim card with MCC262 MNC02.

To make the resources appear again resource de-duping
have to be disabled both when Settings-core is built
and when the Settings APK is built.

Test: make Settings and dump resources with aapt2
aapt2 dump resources Settings.apk
Bug: 171892595

Change-Id: I40318a2ceca33f0d4d6702c6b511420f756b1d34