Don't show irrelevant AppInfo for RRO
RRO packages are displayed in the list of apps in AppInfo.
This is useful for end-user to gauge space taken by them and
make an informed decision on whether or not to remove them if
that is possible. However, there is information that is
not applicable to RRO and can lead to confusion, hence is removed:
- Force stop button:
  RRO cannot be forced stopped like apps
- Open button and "Open by default" preference item:
  RRO cannot be opened like apps
- Notifications preference:
  RRO cannot receive notifications
- Permissions:
  RRO do not request permissions themselves
- Data usage:
  RRO have no code and do not use data
- Battery:
  RRO have no code and do not use battery
- Screen time:
  RRO have no code and one does not spend time in them

Basically the only information that makes sense for RRO within AppInfo
is: Storage, Version and an Uninstall button for non-system RRO.

Bug: 124540315
Test: manual + relevant robotests
Change-Id: I0438b4afcbda806702fe7bdbcc2a2c3816e5a4a0