[adb-wireless] Allow ADB-over-WiFi when WiFi is not default
When other network types with higher priority than wifi is
connected alongside wifi, turning on ADB over wifi is not
possible.

This is happening due to `isWifiConnected()` only checks
if wifi is the default network and not if wifi is connected.

Often times multiple networks such as ethernet is used in
conjuction with wifi, making ADB over wifi not working with
the setup.

Fix by returning if any available network is a WiFi network.

Bug: 173466085
Test: Tested on eve Chrome OS ARCVM, ADB WiFi can be toggled when
      WiFi is not the default network.

Change-Id: I8474a4e8b37a4af0031e1462645509f1e634e9a8