Replace hidden CM#isNetworkSupported(int) usage
It's a refactor work for connectivity mainline module. The hidden
methods access is no longer allowed, so the usage for
isNetworkSupported() should be replaced. Settings use it to check
if device support telephony, wifi or ethernet service. There are
alternative methods to check if device supports such features.
Replace as it is.

Bug: 172183305
Test: m
Change-Id: I8f1d11558b1be575a0777ed195abe027e838cb74